+++ spec/reputation_system/query_methods_spec.rb
@@ -126,7 +126,7 @@ describe ActiveRecord::Base do
       end
     end
   end
-=begin
+
   describe "#with_normalized_reputation" do
     context "Without Scopes" do
       before :each do
@@ -136,30 +136,30 @@ describe ActiveRecord::Base do
       it "should return result with given normalized reputation" do
         @question2 = Question.create!(:text => 'Does this work?', :author_id => @user.id)
         @question2.add_evaluation(:total_votes, 6, @user)
-        res = Question.with_normalized_reputation(:total_votes, :all, {})
+        res = Question.with_normalized_reputation(:total_votes)
         res.should == [@question, @question2]
         res[0].normalized_total_votes.should be_within(DELTA).of(0)
         res[1].normalized_total_votes.should be_within(DELTA).of(1)
       end
 
-      it "should retain select option" do
-        res = Question.with_normalized_reputation(:total_votes, :all, {:select => "questions.id"})
+      it "should not retain select option" do
+        res = Question.with_normalized_reputation(:total_votes).select("questions.id")
         res.should == [@question]
         res[0].id.should_not be_nil
-        lambda {res[0].text}.should raise_error
+        lambda {res[0].text}.should_not raise_error
       end
 
       it "should retain conditions option" do
         @question2 = Question.create!(:text => 'Does this work?', :author_id => @user.id)
         @question2.add_evaluation(:total_votes, 6, @user)
-        res = Question.with_normalized_reputation(:total_votes, :all, {:conditions => "normalized_total_votes > 0.6"})
+        res = Question.with_normalized_reputation(:total_votes).where("normalized_total_votes > 0.6")
         res.should == [@question2]
       end
 
       it "should retain joins option" do
-        res = Question.with_normalized_reputation(:total_votes, :all, {
-          :select => "questions.*, users.name AS user_name",
-          :joins => "JOIN users ON questions.author_id = users.id"})
+        res = Question.with_normalized_reputation(:total_votes).
+          select("questions.*, users.name AS user_name").
+          joins("JOIN users ON questions.author_id = users.id")
         res.should == [@question]
         res[0].user_name.should == @user.name
       end
@@ -174,11 +174,65 @@ describe ActiveRecord::Base do
       end
 
       it "should return result with given reputation" do
-        res = Phrase.with_normalized_reputation(:maturity, :ja, :all, {})
+        res = Phrase.with_normalized_reputation(:maturity, :ja)
         res.should == [@phrase]
         res[0].normalized_maturity.should be_within(DELTA).of(0)
       end
     end
   end
-=end
+
+  describe "#with_normalized_reputation_only" do
+    context "Without Scopes" do
+      before :each do
+        @question.add_evaluation(:total_votes, 3, @user)
+      end
+
+      it "should return result with given normalized reputation" do
+        @question2 = Question.create!(:text => 'Does this work?', :author_id => @user.id)
+        @question2.add_evaluation(:total_votes, 6, @user)
+        res = Question.with_normalized_reputation_only(:total_votes)
+        res.length.should == 2
+        res[0].normalized_total_votes.should be_within(DELTA).of(0)
+        res[1].normalized_total_votes.should be_within(DELTA).of(1)
+      end
+
+      it "should not retain select option" do
+        res = Question.with_normalized_reputation_only(:total_votes).select("questions.id")
+        res.length.should == 1
+        res[0].id.should_not be_nil
+        lambda {res[0].text}.should raise_error
+      end
+
+      it "should retain conditions option" do
+        @question2 = Question.create!(:text => 'Does this work?', :author_id => @user.id)
+        @question2.add_evaluation(:total_votes, 6, @user)
+        res = Question.with_normalized_reputation_only(:total_votes).where("normalized_total_votes > 0.6")
+        res.length.should == 1
+        res[0].normalized_total_votes.should > 0.6
+      end
+
+      it "should retain joins option" do
+        res = Question.with_normalized_reputation_only(:total_votes).
+          select("questions.*, users.name AS user_name").
+          joins("JOIN users ON questions.author_id = users.id")
+        res.length.should == 1
+        res[0].user_name.should == @user.name
+      end
+    end
+
+    context "With Scopes" do
+      before :each do
+        @trans_ja = Translation.create!(:text => "Ichi", :user => @user, :locale => "ja", :phrase => @phrase)
+        @trans_ja.add_evaluation(:votes, 3, @user)
+        @trans_fr = Translation.create!(:text => "Ichi", :user => @user, :locale => "fr", :phrase => @phrase)
+        @trans_fr.add_evaluation(:votes, 6, @user)
+      end
+
+      it "should return result with given reputation" do
+        res = Phrase.with_normalized_reputation_only(:maturity, :ja)
+        res.length.should == 1
+        res[0].normalized_maturity.should be_within(DELTA).of(0)
+      end
+    end
+  end
 end